+++ app/models/vote.rb
@@ -1,0 +1,8 @@
+class Vote < ActiveRecord::Base
+  # Concerns
+  include Deletable
+
+  # Model Relationships
+  belongs_to :user
+  belongs_to :research_topic
+end